update `.type` casts on base augmentation class
- remove all casts around the bases augmentations classes
- add cast on `transform_inputs` of `_AugmentationBase` when autocast region is enabled

+++ kornia/augmentation/_2d/base.py
@@ -1,9 +1,10 @@
 from typing import Any, Dict, Optional
 
-from torch import Tensor, float16, float32, float64
+from torch import float16, float32, float64
 
 from kornia.augmentation.base import _AugmentationBase
 from kornia.augmentation.utils import _transform_input, _validate_input_dtype
+from kornia.core import Tensor
 from kornia.geometry.boxes import Boxes
 from kornia.geometry.keypoints import Keypoints
 from kornia.utils import eye_like
@@ -69,27 +70,20 @@ class RigidAffineAugmentationBase2D(AugmentationBase2D):
 
     def generate_transformation_matrix(self, input: Tensor, params: Dict[str, Tensor], flags: Dict[str, Any]) -> Tensor:
         """Generate transformation matrices with the given input and param settings."""
-        # Note about the explicit type conversions here:
-        # If this code is run in an autocast-enabled region, the transformation matrix or the output
-        # may be float16 even though the input was float32 (e.g. torch.mm produces float16 output),
-        # see also the documentation on autocasting: https://pytorch.org/docs/stable/amp.html.
-        # It may be unexpected for the user if the output type changes and it can also lead to errors
-        # for the index_put operation below (see also https://github.com/kornia/kornia/issues/1737)
-        # In case the type already matches the input type, the conversions are no-ops
-
         to_apply = params['batch_prob']
         in_tensor = self.transform_tensor(input)
         if not to_apply.any():
             trans_matrix = self.identity_matrix(in_tensor)
         elif to_apply.all():
-            trans_matrix = self.compute_transformation(in_tensor, params=params, flags=flags).type(in_tensor.dtype)
+            trans_matrix = self.compute_transformation(in_tensor, params=params, flags=flags)
         else:
             trans_matrix = self.identity_matrix(in_tensor)
+
             trans_matrix = trans_matrix.index_put(
-                (to_apply,),
-                self.compute_transformation(in_tensor[to_apply], params=params, flags=flags).type(in_tensor.dtype),
+                (to_apply,), self.compute_transformation(in_tensor[to_apply], params=params, flags=flags)
             )
-        return trans_matrix.type(in_tensor.dtype)
+
+        return trans_matrix
 
     def inverse_inputs(
         self, input: Tensor, params: Dict[str, Tensor], flags: Dict[str, Any], transform: Optional[Tensor] = None
